+++ miden-tx/src/executor/mast_store.rs
@@ -3,6 +3,7 @@ use core::cell::RefCell;
 
 use miden_lib::{transaction::TransactionKernel, MidenLib, StdLibrary};
 use miden_objects::{
+    accounts::AccountCode,
     assembly::mast::MastForest,
     transaction::{TransactionArgs, TransactionInputs},
     Digest,
@@ -36,18 +37,23 @@ impl TransactionMastStore {
         let store = Self { mast_forests };
 
         // load transaction kernel MAST forest
-        let kernels_forest = Arc::new(TransactionKernel::kernel().into());
+        let kernels_forest = TransactionKernel::kernel().mast_forest().clone();
         store.insert(kernels_forest);
 
         // load miden-stdlib MAST forest
-        let miden_stdlib_forest = Arc::new(StdLibrary::default().into());
+        let miden_stdlib_forest = StdLibrary::default().mast_forest().clone();
         store.insert(miden_stdlib_forest);
 
         // load miden lib MAST forest
-        let miden_lib_forest = Arc::new(MidenLib::default().into());
+        let miden_lib_forest = MidenLib::default().mast_forest().clone();
         store.insert(miden_lib_forest);
 
         store
+    }
+
+    /// Loads the provided account code into this store.
+    pub fn load_account_code(&self, code: &AccountCode) {
+        self.insert(code.mast().clone());
     }
 
     /// Loads code required for executing a transaction with the specified inputs and args into
@@ -59,7 +65,7 @@ impl TransactionMastStore {
     /// - Transaction script (if any) from the specified [TransactionArgs].
     pub fn load_transaction_code(&self, tx_inputs: &TransactionInputs, tx_args: &TransactionArgs) {
         // load account code
-        self.insert(tx_inputs.account().code().mast().clone());
+        self.load_account_code(tx_inputs.account().code());
 
         // load note script MAST into the MAST store
         for note in tx_inputs.input_notes() {