+++ objects/src/transaction/inputs.rs
@@ -70,12 +70,7 @@ impl TransactionInputs {
                         .ok_or(TransactionInputError::InputNoteBlockNotInChainMmr(note.id()))?
                 };
 
-                if !is_in_block(note, proof, block_header) {
-                    return Err(TransactionInputError::InputNoteNotInBlock(
-                        note.id(),
-                        note_block_num,
-                    ));
-                }
+                validate_is_in_block(note, proof, block_header)?;
             }
         }
 
@@ -397,11 +392,20 @@ impl InputNote {
     }
 }
 
-/// Returns true if this note belongs to the note tree of the specified block.
-fn is_in_block(note: &Note, proof: &NoteInclusionProof, block_header: &BlockHeader) -> bool {
+/// Validates whether the provided note belongs to the note tree of the specified block.
+fn validate_is_in_block(
+    note: &Note,
+    proof: &NoteInclusionProof,
+    block_header: &BlockHeader,
+) -> Result<(), TransactionInputError> {
     let note_index = proof.location().node_index_in_block().into();
     let note_hash = note.hash();
-    proof.note_path().verify(note_index, note_hash, &block_header.note_root())
+    proof
+        .note_path()
+        .verify(note_index, note_hash, &block_header.note_root())
+        .map_err(|_| {
+            TransactionInputError::InputNoteNotInBlock(note.id(), proof.location().block_num())
+        })
 }
 
 impl ToInputNoteCommitments for InputNote {